Add support for linking against MPI libraries with threads.
Note that PLATFORM doesn't really work as it outputs the message:

  MPI_File_open is not yet thread safe

so cannot be used with a parallel HDF.

Improve configuration for the METIS library.

Only apply the sanitizer flags for a real GCC, not those who
support GNU extensions.

Tidy up examples Makefile.am to remove compiler specific flags.


Former-commit-id: 8cac59db8fc27033c9f83c13e9d037b757ae64d2