Update task dump plotting scripts.
These fix a bug in the range of adjacent tasks of the same type
(and colour) which meant they merged into a single bar, regardless
of whether there was a gap or not (feature of span_where).

Also update to offer proper command-line options, the main one of
which is --expand. That splits the tasks of a thread into "expand"
lines of tasks, designed so that adjacent tasks will not be on the same
line. It is now also possible to set the size of the output graphic.

The colours have also been tweaked once more.