Now when output_list (snapshots, stf and stats) is turned on, you only get outputs at those times (no first and last dumps).
Special case of when you have an output at time_max is dealt with by flagging
that you need an output in the output_list struct, which is then checked at the
end on the last timestep (in main.c).

Also added a check where you cannot run with Snapshots: invoke_stf and
StructureFinder: output_list at the same time (so we dont hit scenarios where
we could run stf twice).