Inefficient fix for box wrap bug when refining replication list
For cells at the edge of the grid, we check periodic neighbours
too. This fixes the problem but also loses most of the performance
increase. Might be better to wrap particles to centre of cell
rather than into the box and allow an extra half cell boundary
around box replications.