Check for lightcone crossing separately for different types
Previously lightcone crossing check was only done for gparts then
we looked up the type. This is a problem because the part/spart/
bpart may be being updated at the same time. Instead, we now
only do the lightcone check for DM or neutrino gparts and do
a separate check for parts, sparts and bparts.