If flux would reduce a cell to 0 mass or less, do not do flux interaction.
This change prevents serious damage to code, causing issues in chemistry, cooling, gravity, and was often related to the presence of NaNs. Further testing for better solutions to be done, this may end up being the best solution. If this still exists as 0, it is because no better solution was found.